fix: make the state variables static (#247)
* fix: make autoRecordScreenViews variable static

This is needed to properly track the Screen recording and ensure correct state is present, even when class goes through the attach/detach cycle.

* fix: refactor ActivityLifeCycleManager to use static variables for activity count and background state

This is needed to ensure that Application Opened and session tracking features works as expected and ensure correct state is present, even when class goes through the attach/detach cycle.