fix: npe and memory leak issue (#245)
* fix: ActivityLifecycleCallbacks duplicate registration causing multiple event triggers

* refactor: change constructor visibility to package-private in ActivityLifeCycleManager

* fix: npe and memory leak issue

* fix: clear lifecycle events list to prevent memory leak

* fix: NPE issue caused due to lifecycle refactor

If app is moved to background using back button, even before SDK init and opened again and SDK init happens -> as of now it triggers NPE. Now, I've made an optimisation to fix the issue from root.

* fix: add null check for plugin in executeRunnableLifeCycleEvent method

* chore: add @NonNull annotation to plugin parameter in RunnableLifeCycleEventsInterface methods

* fix: make the state variables static (#247)

* fix: make autoRecordScreenViews variable static

This is needed to properly track the Screen recording and ensure correct state is present, even when class goes through the attach/detach cycle.

* fix: refactor ActivityLifeCycleManager to use static variables for activity count and background state

This is needed to ensure that Application Opened and session tracking features works as expected and ensure correct state is present, even when class goes through the attach/detach cycle.